+++ CodeWars/Fundamentals/DuplicateEncoder/Java/src/DuplicateEncoderTest.java
@@ -1,0 +1,17 @@
+/**
+ * Created by cajetan on 21/07/2017.
+ */
+
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+import org.junit.runners.JUnit4;
+
+
+public class DuplicateEncoderTest {
+    @Test
+    public void test() {
+        assertEquals(")()())()(()()(",
+                DuplicateEncoder.encode("Prespecialized"));
+        assertEquals("))))())))",DuplicateEncoder.encode("   ()(   "));
+    }
+}